Removing e dir warning messages.

+++ M2/Macaulay2/e/monoid.hpp
@@ -25,10 +25,10 @@ typedef const int * const_monomial;
 typedef const int * const_varpower;
 
 #define ALLOCATE_EXPONENTS(byte_len) static_cast<exponents>(alloca(byte_len))
-#define EXPONENT_BYTE_SIZE(nvars) (sizeof(int) * (nvars))
+#define EXPONENT_BYTE_SIZE(nvars) static_cast<int>((sizeof(int) * (nvars)))
 
 #define ALLOCATE_MONOMIAL(byte_len) static_cast<monomial>(alloca(byte_len))
-#define MONOMIAL_BYTE_SIZE(mon_size) (sizeof(int) * (mon_size))
+#define MONOMIAL_BYTE_SIZE(mon_size) static_cast<int>((sizeof(int) * (mon_size)))
 
 class Monoid : public mutable_object
 {
